fix: Ack endpoint discovery
Previously, the xDS client did not receive updates after the first
response. This could cause problems if the endpoint resources change, or
if they are not present in the initial response, e.g. due to a race.

The problem is that the xDS server loop waits until it receives a
DiscoveryRequest before sending a DiscoveryResponse.

This change fixes the issue by sending a DiscoveryRequest after each
response is received, updating it with the last seen version and
response nonce.

This change also improves unit test coverage, including testing for the
issue being addressed here.

Fixes: #16